m16_add_django/pyproject.toml
[tool.poetry]
name = "m16-add-django"
version = "0.1.0"
description = ""
authors = ["Yuriy Kuchma <krabatua@gmail.com>"]
readme = "README.md"
packages = [{include = "m16_add_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
celery = "^5.3.6"
pymongo = "^4.6.1"
redis = "^5.0.1"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
python-dateutil = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
